APIBackend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
djangorestframework==3.14.0
install==1.3.5
pytz==2022.7
sqlparse==0.4.3
